mod: create seperate module of the wifi creadentials
